Move AccountController to controller package
