+++ src/main/java/georgemarrows/learnspring/controller/DemoApplication.java
@@ -1,4 +1,4 @@
-package georgemarrows.learnspring.demo;
+package georgemarrows.learnspring.controller;
 
 import java.util.Arrays;
