+++ src/test/java/georgemarrows/learnspring/controller/AccountControllerIntegrationTest.java
@@ -1,15 +1,10 @@
-package georgemarrows.learnspring.demo;
+package georgemarrows.learnspring.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.json.JSONObject;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.logging.LogLevel;
-import org.springframework.boot.logging.LoggingSystem;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.HttpEntity;
@@ -19,13 +14,6 @@ import org.springframework.http.ResponseEntity;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class AccountControllerIntegrationTest {
-
-	@BeforeAll
-	public static void setErrorLogging() {
-	   LoggingSystem.get(ClassLoader.getSystemClassLoader()).setLogLevel(Logger.ROOT_LOGGER_NAME, LogLevel.DEBUG);
-	}
-
-	Logger logger = LoggerFactory.getLogger(AccountController.class);
 
 	@Autowired
 	private TestRestTemplate template;
